Reusable components for analysis pages and firebase events
