+++ src/components/analysis/JobListItem.tsx
