+++ src/components/analysis/ld50/AnalysisSetupPanel.tsx
