+++ src/components/analysis/ld50/DataInput.tsx
