+++ src/components/analysis/nmr/VarianDataInput.tsx
